CLI-837 Move migration.ts into core/update/ alongside claude-hooks-migration.ts
commands/integrate/claude/index.ts calling removeObsoleteHookArtifacts
directly isn't a separate concern from the post-update flow -- it's the same
obsolete-a3s-hook cleanup, just triggered by a manual integrate re-run instead
of an automatic upgrade. Grouping by purpose (this is all Claude-hooks
migration code) is more coherent than grouping by "who calls it directly",
especially since core/update/ already exists as the right home and
migration.ts has no host-specific dependency of its own.
