+++ src/commands/integrate/claude/index.ts
@@ -24,8 +24,8 @@ import { homedir } from 'node:os';
 
 import { installIntegration } from '@/core/framework/features';
 import type { ResolvedAuth } from '@/core/host/auth-resolver.ts';
-import { OBSOLETE_A3S_MARKER, removeObsoleteHookArtifacts } from '@/core/host/migration.ts';
 import type { IntegrationStateAttribute } from '@/core/state/state.ts';
+import { OBSOLETE_A3S_MARKER, removeObsoleteHookArtifacts } from '@/core/update/migration.ts';
 
 import { printAgentNonInteractiveAlternativeHint } from '../../_common/agent-prompt-hint.ts';
 import {
