+++ src/core/update/claude-hooks-migration.ts
@@ -25,15 +25,14 @@ import * as fs from 'node:fs';
 import { homedir } from 'node:os';
 import { join } from 'node:path';
 
+import logger from '../observability/logger.ts';
+import type { CliState, HookExtension } from '../state/state.ts';
+import { loadState } from '../state/state-repository.ts';
 import {
   migrateHookScripts,
   OBSOLETE_A3S_MARKER,
   removeObsoleteHookArtifacts,
-} from '@/core/host/migration.ts';
-
-import logger from '../observability/logger.ts';
-import type { CliState, HookExtension } from '../state/state.ts';
-import { loadState } from '../state/state-repository.ts';
+} from './migration.ts';
 
 /**
  * Signature of `@/commands/integrate/claude/hooks.ts`'s `installHooks`,
