+++ tests/unit/core/update/migration.test.ts
@@ -24,15 +24,15 @@ import { join } from 'node:path';
 
 import { afterEach, beforeEach, describe, expect, it } from 'bun:test';
 
+import type { AgentExtension, CliState, HookExtension } from '@/core/state/state.ts';
+import { getDefaultState } from '@/core/state/state.ts';
+import { setMockUi } from '@/core/ui';
 import {
   cleanObsoleteFromState,
   migrateHookScripts,
   OBSOLETE_A3S_MARKER,
   removeObsoleteHookArtifacts,
-} from '@/core/host/migration.ts';
-import type { AgentExtension, CliState, HookExtension } from '@/core/state/state.ts';
-import { getDefaultState } from '@/core/state/state.ts';
-import { setMockUi } from '@/core/ui';
+} from '@/core/update/migration.ts';
 
 const OLD_VERSION = '0.4.0';
