+++ tests/unit/core/update/post-update.test.ts
@@ -26,10 +26,10 @@ import { supportedIntegrations } from '@/commands/integrate';
 import { CLAUDE_INTEGRATION_ID } from '@/commands/integrate/claude/declaration.ts';
 import * as hooks from '@/commands/integrate/claude/hooks.ts';
 import * as secretsInstall from '@/core/host/install/secrets.ts';
-import * as migration from '@/core/host/migration.ts';
 import type { CliState } from '@/core/state/state.ts';
 import { getDefaultState } from '@/core/state/state.ts';
 import * as stateRepository from '@/core/state/state-repository.ts';
+import * as migration from '@/core/update/migration.ts';
 import type { PostUpdateDependencies } from '@/core/update/post-update.ts';
 import { runPostUpdateActions } from '@/core/update/post-update.ts';
 import * as versionLib from '@/core/version.ts';
